[REF][SLOT-153] add refactor
